requirements.txt
Django==2.2.1
django-crispy-forms==1.7.2
Pillow==6.0.0
djangorestframework==3.9.4
